+++ packages/state-manager/src/sagas/appConnection/connection.adapter.ts
@@ -1,10 +1,6 @@
 import { createEntityAdapter } from '@reduxjs/toolkit'
 import { NetworkStats } from './connection.slice'
 
-export const connectedPeersAdapter = createEntityAdapter<string>({
-  selectId: (connectedPeer) => connectedPeer
-})
-
 export const peersStatsAdapter = createEntityAdapter<NetworkStats>({
   selectId: (peer) => peer.peerId
 })